+++ .eslintignore
@@ -1,2 +1,3 @@
 jquery.js
 bootstrap-3.3.5/
+static/libs/
