+++ .eslintrc.json
@@ -1,4 +1,9 @@
 {
+  "parserOptions": {
+    "sourceType": "module",
+    "ecmaVersion": 6
+
+  },
   "env": {
     "browser": true,
     "jquery": true,
