--- dialog_edit.html
@@ -1,50 +1,0 @@
-<div id="bg" class="bg"></div>
-<div id="dialog" class="col-12 col-sm-12 offset-md-2 col-md-8 offset-lg-2 col-lg-8 offset-xl-3 col-xl-6">
-    <div class="row">
-        <div class="col-8 offset-2">
-            <img id="pic" src=""/>
-            <div id="drop_zone">Serien Bild</div>
-        </div>
-        <div class="col-2">
-            <button id="close" class="btn btn-link" type="button">
-                <i class="material-icons">close</i>
-            </button>
-        </div>
-    </div>
-    <form id="form" action="javascript:void(0);">
-        <div id="row-titel" class="row">
-            <div class="offset-sm-2 col-10 col-sm-8">
-                <input id="titel" name="titel" type="text" pattern="[a-zA-Z0-9]+([a-zA-Z0-9 \-]*[a-zA-Z0-9\-])*" required
-                    placeholder="Titel" autocomplete="off" list="titelList">
-            </div>
-            <div class="col-2">
-                <button id="delete" class="btn btn-link" type="button">
-                    <i class="material-icons">delete</i>
-                </button>
-            </div>
-        </div>
-        <div class="row">
-            <div class="col-2">
-                <button id="SUP" class="btn btn-link" type="button">
-                    <i class="material-icons">add</i>
-                </button>
-            </div>
-            <div class="col-8">
-                <input id="stand" name="stand" type="text" pattern="^((S|B)[0-9x]{2}E[0-9x]{2}|E[0-9x]{5})$" required
-                    placeholder="N&auml;chste Episode" autocomplete="off">
-            </div>
-            <div class="col-2">
-                <button id="EUP" class="btn btn-link" type="button">
-                    <i class="material-icons">add</i>
-                </button>
-            </div>
-        </div>
-        <div class="row">
-            <div class="offset-md-2 col col-md-8">
-                <button id="submit" class="btn btn-link" type="button">
-                    <i class="material-icons">done</i>
-                </button>
-            </div>
-        </div>
-    </form>
-</div>
