--- dialog_error.html
@@ -1,7 +1,0 @@
-<div id="bg" class="bg"></div>
-<div id="dialog">
-    <div class="d-flex align-items-end flex-column" style="height: 200px; position:fixed; top: 110px; right: 20px; z-index: 5;">
-        <div id="error-dialog" class="p-5 dialog" style="text-align: left">
-        </div>
-    </div>
-</div>
