--- dialog_info.html
@@ -1,17 +1,0 @@
-<div id="bg" class="bg"></div>
-<div class="d-flex align-items-end flex-column p-5" >
-    <div id="dialog" style="text-align: left">
-        Style:<br>
-        <a href="#" id="changestyle">Switch Style</a><br><br>
-        Stuff used:<br>
-        Google Material Icons: <a target="_blank" href="https://material.io/icons/">https://material.io/icons/</a><br>
-        Bootstrap: <a target="_blank" href="https://v4-alpha.getbootstrap.com/">https://v4-alpha.getbootstrap.com/</a><br>
-        JQuery: <a target="_blank" href="https://jquery.com/">https://jquery.com/</a><br>
-        JQuery Lazy: <a target="_blank" href="http://jquery.eisbehr.de/lazy/">http://jquery.eisbehr.de/lazy/</a><br>
-        <br>
-        Download status: <a href="server/get_status.php">status.txt</a><br/>
-        Upload status: <form id="upload_status" action="server/post_status_file.php" method="post" enctype="multipart/form-data"><input type="file" name="file" id="file"><input class="submit" type="submit" value="Upload" name="submit"></form>
-        <br>
-        Made by <a href="https://github.com/Sandr00">Sandr0</a>
-    <!-- </div> -->
-</div>
