+++ sass/series.scss
@@ -1,0 +1,178 @@
+* {
+  text-align: center;
+  text-decoration: none;
+  padding: 0;
+  margin: 0;
+  outline: none;
+  border: 0 solid #333;
+}
+
+.container-fluid {
+  margin: 0;
+  padding: 0;
+}
+
+body {
+  /*font-size: 0;*/
+}
+
+.row div, #nav div {
+  padding: 0;
+}
+
+.shadow {
+  left: 0;
+  bottom: 0;
+  z-index: 2;
+  position: absolute;
+  width: 100%;
+  font-size: 1.6em;
+  background-size: 100% 100%;
+  padding-bottom: 2%;
+  padding-top: 20%;
+}
+
+.shadow {
+  box-shadow: none !important;
+}
+
+.shadow.x {
+  text-decoration: line-through;
+}
+
+.series {
+  position: absolute;
+  top: 0;
+  left: 0;
+  bottom: 0;
+  right: 0;
+  background: url(../img/unknown.jpg) center, no-repeat;
+  background-size: cover;
+  margin: 4px;
+  width: auto;
+}
+
+.seriesDiv:before{
+  content: "";
+  display: block;
+  padding-top: 165%; /* Ratio */
+}
+
+#overlay {
+  font-size: 16px;
+  position: absolute;
+  top: 0;
+  padding: 0;
+}
+
+.bg, #dialog {
+  position: fixed;
+}
+
+.bg {
+  z-index: 4;
+  width: 100%;
+  height: 100%;
+  top: 0;
+  left: 0;
+  background: #000;
+  opacity: 0.8;
+}
+
+#dialog {
+  z-index: 5;
+  top: 100px;
+  padding: 20px;
+  border: 2px solid #333;
+  border-radius: 4px;
+  min-width: 320px;
+  box-shadow: 0 0 10px 0 rgba(0, 0, 0, .5);
+}
+
+#pic, #titel, #drop_zone {
+  border-width: 2px;
+}
+
+#row-titel {
+  margin-top: 20px;
+}
+
+#pic, #drop_zone {
+  height: 300px;
+}
+
+#drop_zone {
+  border: dashed #bbb;
+  -moz-border-radius: 5px;
+  -webkit-border-radius: 5px;
+  border-radius: 5px;
+  padding: 25px;
+  text-align: center;
+  color: #bbb;
+  min-width: 200px;
+}
+
+input, .input, #submit {
+  width:100%
+  /*background: #FFF;*/
+}
+
+input, #EUP, #SUP, .btn {
+  line-height: 42px;
+  height: 42px;
+  margin: 2px;
+}
+
+
+.navbar {
+  min-height: 42px !important;
+  z-index: 5;
+  border-radius: 0 0 0 0;
+  box-shadow: 0 0 10px 0 rgba(0, 0, 0, 0.5);
+}
+
+.nav-item {
+  margin-left: 10px;
+  margin-right: 10px;
+}
+
+a:hover {
+  cursor: pointer;
+}
+
+/* Tooltip */
+.tooltip-inner {
+  font-size: 1.4em;
+}
+.tooltip {
+  pointer-events: none;
+}
+
+/* Refresh circle */
+#refresh {
+  cursor: pointer;
+  height: auto;
+  transform: rotate(-90deg);
+  width: 60px;
+}
+#refresh circle.incomplete {
+  opacity: 0.25;
+}
+#refresh circle.complete {
+  stroke-dasharray: 100px;
+  transition: stroke-dashoffset 10000ms ease-in-out;
+}
+#refresh circle {
+  fill: rgba(0,0,0,0);
+  stroke-width: 2;
+}
+#upload_status input {
+  display: inline-block;
+  width: auto;
+  background: none;
+  padding: 2px;
+}
+#upload_status input.submit {
+  padding-left: 10px;
+  padding-right: 10px;
+}
