+++ static/test.html
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="content-type" content="text/html;"/>
+    <link type='text/css' href="css/bootstrap.min.css" rel="stylesheet">
+    <!-- My Stylesheet -->
+    <link type="text/css" href="css/series.css" rel="stylesheet">
+    <link type="text/css" href="css/color.css" rel="stylesheet">
+    <title>Serien</title>
+</head>
+<body class="dark">
+<single-series id="seriesContent" class="container-fluid">
+</single-series>
+
+<script type="module" src="js/dialogComp.js"></script>
+<script type="module" src="js/comp.js"></script>
+
+</body>
+</html>
